Prometheus history: combine data for the same pod
The raw data in Prometheus is usually collected at cgroup level, so the
same pod/container can appear multiple times, throwing off the
processing logic.

For example this will happen when a statefulset pod restarts: it retains
the same pod identity and container name, but instantiated by a new
container with a new cgroup identity.

Using `sum()` for CPU gives valid results, because `rate` smooths the
contributions of old and new series. However we use `max()` for memory
to avoid double-counting the old and new.

As a bonus, `sum()` allows this code to work when cAdvisor is configured
to report per-core CPU consumption, which is an option.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>