Prometheus history: skip pods that lack the expected label
The code is looking for a label used in selectors, like `app=foo`.
But it is ok to have other pods not used in selectors; we should
just ignore those.

Still return an error if we went all through the pods and didn't find
a single one with the right label; that could mean the config is wrong.

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>